docs(sdk): tighten doc + safety around push_raw_message, payload size, variant extension, fetcher type
- push_raw_message: clarify it's the eager-only path; plugins needing lazy
  materialization or ObjectIngestPolicy dispatch use push_message_v2.
- PJ_payload_t::size: size_t -> uint64_t for ABI-explicit 64-bit width
  (the protocol pins everything else to 64-bit; staying with size_t was
  a portability hazard on cross-compiles).
- CanonicalObject variant: rewrite the extension-policy comment. Appending
  alternatives is forward-compatible — older hosts receiving an unknown
  kind reject the message, no protocol bump required.
- pushMessage template: static_assert that the Fetcher returns either
  sdk::PayloadView or std::vector<uint8_t>. Catches misuse at compile
  time instead of falling through to UB in the else branch.

+++ pj_base/include/pj_base/data_source_protocol.h
@@ -152,7 +152,7 @@ typedef struct PJ_payload_anchor_t {
  */
 typedef struct PJ_payload_t {
   const uint8_t* data;
-  size_t size;
+  uint64_t size;
   PJ_payload_anchor_t anchor;
 } PJ_payload_t;
 
@@ -254,6 +254,12 @@ typedef struct PJ_data_source_runtime_host_vtable_t {
    * @p handle must have been obtained from ensure_parser_binding.
    * @p host_timestamp_ns is nanoseconds since the Unix epoch
    * (1970-01-01T00:00:00Z). Returns false + error on failure.
+   *
+   * Eager-only push: the host parses immediately and the bytes are not
+   * retained for later replay. Plugins that need lazy materialization or
+   * ObjectIngestPolicy dispatch should use push_message_v2 instead. This
+   * slot remains for sources that fan-out raw bytes without an associated
+   * fetcher (streaming or eager-only consumers).
    */
   bool (*push_raw_message)(
       void* ctx, PJ_parser_binding_handle_t handle, int64_t host_timestamp_ns, PJ_bytes_view_t payload,
